requirements.txt
click==8.0.4
colorama==0.4.4
Pillow==9.0.1
sty==1.0.4
